requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
imblearn
pytorch
xgboost
scipy
seaborn
matplotlib
